TCP-LLM-SERVER/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
choreographer==1.0.10
exceptiongroup==1.3.0
google-auth==2.40.3
google-genai==1.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.2
kaleido==1.0.0
logistro==1.1.0
narwhals==2.3.0
numpy==2.2.6
orjson==3.11.3
packaging==25.0
pillow==11.3.0
plotly==6.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
simplejson==3.20.1
sniffio==1.3.1
tenacity==9.1.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
